modify codegen
