--- client/graphql/gql.ts
@@ -1,58 +1,0 @@
-/* eslint-disable */
-import * as types from './graphql';
-import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/core';
-
-/**
- * Map of all GraphQL operations in the project.
- *
- * This map has several performance disadvantages:
- * 1. It is not tree-shakeable, so it will include all operations in the project.
- * 2. It is not minifiable, so the string of a GraphQL query will be multiple times inside the bundle.
- * 3. It does not support dead code elimination, so it will add unused operations.
- *
- * Therefore it is highly recommended to use the babel or swc plugin for production.
- * Learn more about it here: https://the-guild.dev/graphql/codegen/plugins/presets/preset-client#reducing-bundle-size
- */
-type Documents = {
-    "\n  query CurrentUser {\n    currentUser {\n      id\n      uid\n      createdAt\n      updatedAt\n      profile {\n        id\n        name\n        birthDate\n        gender\n        height\n        weight\n        activityLevel\n      }\n    }\n  }\n": typeof types.CurrentUserDocument,
-    "\n  query ProfileEditCurrentUser {\n    currentUser {\n      id\n      uid\n      profile {\n        id\n        name\n        birthDate\n        gender\n        height\n        weight\n        activityLevel\n        imageURL\n      }\n    }\n  }\n": typeof types.ProfileEditCurrentUserDocument,
-    "\n  mutation UpdateProfile($input: UpdateProfile!) {\n    updateProfile(input: $input) {\n      id\n      name\n      birthDate\n      gender\n      height\n      weight\n      activityLevel\n      imageURL\n    }\n  }\n": typeof types.UpdateProfileDocument,
-};
-const documents: Documents = {
-    "\n  query CurrentUser {\n    currentUser {\n      id\n      uid\n      createdAt\n      updatedAt\n      profile {\n        id\n        name\n        birthDate\n        gender\n        height\n        weight\n        activityLevel\n      }\n    }\n  }\n": types.CurrentUserDocument,
-    "\n  query ProfileEditCurrentUser {\n    currentUser {\n      id\n      uid\n      profile {\n        id\n        name\n        birthDate\n        gender\n        height\n        weight\n        activityLevel\n        imageURL\n      }\n    }\n  }\n": types.ProfileEditCurrentUserDocument,
-    "\n  mutation UpdateProfile($input: UpdateProfile!) {\n    updateProfile(input: $input) {\n      id\n      name\n      birthDate\n      gender\n      height\n      weight\n      activityLevel\n      imageURL\n    }\n  }\n": types.UpdateProfileDocument,
-};
-
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- *
- *
- * @example
- * ```ts
- * const query = graphql(`query GetUser($id: ID!) { user(id: $id) { name } }`);
- * ```
- *
- * The query argument is unknown!
- * Please regenerate the types.
- */
-export function graphql(source: string): unknown;
-
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query CurrentUser {\n    currentUser {\n      id\n      uid\n      createdAt\n      updatedAt\n      profile {\n        id\n        name\n        birthDate\n        gender\n        height\n        weight\n        activityLevel\n      }\n    }\n  }\n"): (typeof documents)["\n  query CurrentUser {\n    currentUser {\n      id\n      uid\n      createdAt\n      updatedAt\n      profile {\n        id\n        name\n        birthDate\n        gender\n        height\n        weight\n        activityLevel\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query ProfileEditCurrentUser {\n    currentUser {\n      id\n      uid\n      profile {\n        id\n        name\n        birthDate\n        gender\n        height\n        weight\n        activityLevel\n        imageURL\n      }\n    }\n  }\n"): (typeof documents)["\n  query ProfileEditCurrentUser {\n    currentUser {\n      id\n      uid\n      profile {\n        id\n        name\n        birthDate\n        gender\n        height\n        weight\n        activityLevel\n        imageURL\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation UpdateProfile($input: UpdateProfile!) {\n    updateProfile(input: $input) {\n      id\n      name\n      birthDate\n      gender\n      height\n      weight\n      activityLevel\n      imageURL\n    }\n  }\n"): (typeof documents)["\n  mutation UpdateProfile($input: UpdateProfile!) {\n    updateProfile(input: $input) {\n      id\n      name\n      birthDate\n      gender\n      height\n      weight\n      activityLevel\n      imageURL\n    }\n  }\n"];
-
-export function graphql(source: string) {
-  return (documents as any)[source] ?? {};
-}
-
-export type DocumentType<TDocumentNode extends DocumentNode<any, any>> = TDocumentNode extends DocumentNode<  infer TType,  any>  ? TType  : never;
