+++ client/package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@apollo/client": "^3.13.8",
     "@expo/vector-icons": "^14.1.0",
+    "@graphql-codegen/typescript-graphql-request": "^6.3.0",
     "@react-native-firebase/app": "^22.4.0",
     "@react-native-firebase/auth": "^22.4.0",
     "@react-native-firebase/crashlytics": "^22.4.0",
